bgpv1: Consolidate CRD field types to follow K8s API Conventions
Modifies the types of CRD fileds to follow  the K8s API
Conventions rules:
- "Use pointer types for all optional fields that do not have
built-in nil value"
- "Do not use unsigned integers, due to inconsistent support
across languages and libraries"
- "All public integer fields MUST use the Go (u)int32 or Go
 (u)int64 types, not (u)int"

Also changes the types of the existing ASN CRD fileds to int64
to satisfy the above rules.

Defaults ExportPodCIDR field of CiliumBGPVirtualRouter to false
explicitly for consistency with other optional fields.
As the default behavior remains the same, this is
a backward-compatible change.

Pointer types can be directly dereferenced in the reconcilers code.
To allow that we provide defaulting methods for our API types,
that can be used from the controllers / unit tests.

+++ pkg/bgpv1/agent/annotations_test.go
@@ -21,7 +21,7 @@ func TestAnnotation(t *testing.T) {
 		// expected parsed attributes
 		attr Attributes
 		// expected parsed asn
-		asn int
+		asn int64
 		// error nil or not
 		error error
 	}{
